lessons/lesson_15_pytest_2/requirements.txt
pytest
pytest-xdist
allure-pytest
filelock
